Relocate components into new folder
Now that we have separate components for everything, move them all into their own folders so they can have their own stylesheets, etc.
