+++ src/components/App/styles.css
@@ -1,0 +1,3 @@
+.App-container {
+  text-align: center;
+}
