+++ src/components/Header/styles.css
@@ -1,24 +1,21 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 80px;
-}
-
-.App-header {
+.Header-container {
   background-color: #222;
   height: 150px;
   padding: 20px;
   color: white;
 }
 
-.App-intro {
-  font-size: large;
+.Header-logo {
+  animation: logo-spin infinite 20s linear;
+  height: 80px;
 }
 
-@keyframes App-logo-spin {
+.Header-intro {
+  font-size: large;
+  padding: 1rem 2rem;
+}
+
+@keyframes Header-logo-spin {
   from { transform: rotate(0deg); }
   to { transform: rotate(360deg); }
 }
