+++ src/index.js
@@ -3,7 +3,7 @@ dotenv.config()
 
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './App'
+import App from './components/App'
 import './index.css'
 
 ReactDOM.render(
